Adding k2sStart and Stop as a seperate test file scenario
